refactor: extract all that needed to be imported into single index file
