+++ src/features/user-profile/index.ts
@@ -1,2 +1,2 @@
 export { useUserProfile } from "./hooks/useUserProfile";
-export type { UserProfileResponse } from "./model/type";
+export type { UserProfileResponse } from "./model/types";
